+++ lib/paper_trail/version_number.rb
@@ -7,7 +7,7 @@ module PaperTrail
   module VERSION
     MAJOR = 7
     MINOR = 0
-    TINY = 0
+    TINY = 1
     PRE = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".").freeze